+++ fp-bindgen/src/prelude.rs
@@ -2,9 +2,12 @@ pub use crate::functions::{Function, FunctionList};
 pub use crate::primitives::Primitive;
 pub use crate::serializable::Serializable;
 pub use crate::types::{CustomType, Type, TypeIdent, TypeMap};
-#[cfg(feature = "generators")]
+#[cfg(any(feature = "generators", feature = "wasmtime-core-wasm"))]
 pub use crate::{
-    try_generate_wasmtime_core_wasm_bindings, BindingConfig, BindingsType, RustPluginConfig,
-    RustPluginConfigValue, TsRuntimeConfig, WasmtimeCoreWasmError,
+    generate_bindings, BindingConfig, BindingsType, WasmtimeCoreWasmError,
 };
+#[cfg(feature = "generators")]
+pub use crate::{RustPluginConfig, RustPluginConfigValue, TsRuntimeConfig};
+#[cfg(feature = "wasmtime-core-wasm")]
+pub use crate::try_generate_wasmtime_core_wasm_bindings;
 pub use fp_bindgen_macros::*;